move category.util to dates formatter file
